Minimised file size and reduced redundant code / deleted old classes
New DICOMLoader interface now functional
