+++ src/com/roundsToThree/sd4j.java
@@ -2,20 +2,16 @@ package com.roundsToThree;
 
 import com.roundsToThree.Exception.InvalidFileException;
 import com.roundsToThree.FileIO.DICOMLoader;
-import com.roundsToThree.FileIO.DICOMLoader2;
 import com.roundsToThree.Representations.DateTimeRepresentation;
 import com.roundsToThree.Representations.PersonRepresentation;
 import com.roundsToThree.Representations.Representation;
-import com.roundsToThree.Structures.DataElement;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashMap;
 
 public class sd4j {
-    public HashMap<Integer, DataElement> elements;
-    public HashMap<Integer, Representation> elements2;
+    public HashMap<Integer, Representation> elements;
 
     private sd4j() {
     }
@@ -28,7 +24,7 @@ public class sd4j {
 
         // Read the file
         try {
-            DICOMLoader2.loadDICOMSliceFromFile(simpledicom, sliceFile);
+            DICOMLoader.loadDICOMSliceFromFile(simpledicom, sliceFile);
         } catch (IOException e) {
             System.out.print("Failed to load DICOM file " + sliceFile.getAbsolutePath() + ", IO Exception: ");
             System.out.println(e.getMessage());
@@ -42,22 +38,17 @@ public class sd4j {
     }
 
     // Return the elements given a group number
-    public DataElement[] getElementsByGroupNumber(int groupNumber) {
-        return null;
-    }
+//    public DataElement[] getElementsByGroupNumber(int groupNumber) {
+//        return null;
+//    }
 
     // Returns an element given an group and element number
-    public DataElement getElementByTagNumber(int groupNumber, int elementNumber) {
-        int ind = (int) ((groupNumber << 16) | (elementNumber & 0xFFFF));
-        return elements.get(ind);
+    public Representation getElementByTagNumber(int groupNumber, int elementNumber) {
+        return elements.get((int) ((groupNumber << 16) | (elementNumber & 0xFFFF)));
     }
 
     public PersonRepresentation getPatientName() {
-        DataElement patientName = getElementByTagNumber(0x0010, 0x0010);
-        if (patientName == null || patientName.items == null || patientName.items.size() == 0)
-            return null;
-
-        return new PersonRepresentation(patientName.items.get(0).value);
+        return (PersonRepresentation) getElementByTagNumber(0x0010, 0x0010);
     }
